fix: country and currency seed
